[Chore] 파일 나눠서 배치
