+++ Manito/Manito/Screens/CreateRoom/CreateRoomViewController.swift
@@ -26,37 +26,6 @@ class CreateRoomViewController: BaseViewController {
         return btn
     }()
     
-    // 방 제목입력 뷰
-    private let createView : UIView = {
-        let view = UIView()
-        view.backgroundColor = .backgroundGrey
-        return view
-    }()
-    
-    private let roomsNameTextField: UITextField = {
-        let tf = UITextField()
-        tf.backgroundColor = UIColor.subBackgroundGrey
-        let attributes = [
-            NSAttributedString.Key.foregroundColor : UIColor.white,
-            NSAttributedString.Key.font : UIFont(name: AppFontName.regular.rawValue, size: 18)!
-        ]
-        tf.attributedPlaceholder = NSAttributedString(string: "방 이름을 적어주세요", attributes:attributes)
-        
-        tf.layer.cornerRadius = 10
-        tf.layer.masksToBounds = true
-        tf.layer.borderWidth = 1
-        tf.layer.borderColor = UIColor.lightGray.cgColor
-        tf.textAlignment = .center
-        return tf
-    }()
-    
-    private let roomsTextLimit : UILabel = {
-        let label = UILabel()
-        label.text = "0/8"
-        label.font = .font(.regular, ofSize: 20)
-        return label
-    }()
-    
     lazy var nextButton: UIButton = {
         let btn = UIButton(type: .system)
         btn.setTitle("다음", for: .normal)
@@ -68,94 +37,19 @@ class CreateRoomViewController: BaseViewController {
         return btn
     }()
     
-    // 방 인원수 입력 뷰
-    private let personView : UIView = {
-        let view = UIView()
-        view.backgroundColor = .backgroundGrey
+    private let nameView: InputNameView = {
+        let view = InputNameView()
         return view
     }()
     
-    private let peronsViewLabel: UILabel = {
-        let label = UILabel()
-        label.text = "최대 참여 인원을 설정해 주세요"
-        label.font = .font(.regular, ofSize: 18)
-        return label
-    }()
-    
-    private let personBackView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .darkGray
-        view.layer.borderWidth = 1
-        view.layer.borderColor = UIColor.strokeYello.cgColor
-        view.layer.cornerRadius = 10
+    private let personView: InputPersonView = {
+        let view = InputPersonView()
         return view
     }()
     
-    private let imageView: UIImageView = {
-        let iv = UIImageView()
-        iv.image = UIImage(systemName: "heart")
-        iv.backgroundColor = .darkGray
-        return iv
-    }()
-    
-    private var personLabel: UILabel = {
-        let label = UILabel()
-        label.text = "X 5인"
-        label.font = .font(.regular, ofSize: 24)
-        return label
-    }()
-    
-    lazy var personSlider: UISlider = {
-        let slider = UISlider()
-        slider.value = 1
-        slider.minimumValue = 5
-        slider.maximumValue = 15
-        slider.addTarget(self, action: #selector(didSlider(_:)), for: .valueChanged)
-        return slider
-    }()
-    
-    private var minLabel: UILabel = {
-        let label = UILabel()
-        label.text = "5인"
-        label.font = .font(.regular, ofSize: 16)
-        return label
-    }()
-    
-    private var maxLabel: UILabel = {
-        let label = UILabel()
-        label.text = "15인"
-        label.font = .font(.regular, ofSize: 16)
-        return label
-    }()
-    
-    // 방 기한 설정 뷰
-    private let dateView : UIView = {
-        let view = UIView()
-        view.backgroundColor = .backgroundGrey
+    private let dateView: InputDateView = {
+        let view = InputDateView()
         return view
-    }()
-    
-    private let dateViewLabel: UILabel = {
-        let label = UILabel()
-        label.text = "활동 기간을 설정해 주세요"
-        label.font = .font(.regular, ofSize: 18)
-        return label
-    }()
-    
-    private let dateBackView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .darkGray
-        view.layer.borderWidth = 1
-        view.layer.borderColor = UIColor.strokeYello.cgColor
-        view.layer.cornerRadius = 10
-        return view
-    }()
-    
-    private var dateLabel: UILabel = {
-        let label = UILabel()
-        label.text = "2022.06.06 ~ 2022.06.11"
-        label.font = .font(.regular, ofSize: 20)
-        return label
     }()
     
     // MARK: - Life Cycle
@@ -169,168 +63,58 @@ class CreateRoomViewController: BaseViewController {
         view.addSubview(titleLabel)
         titleLabel.snp.makeConstraints {
             $0.top.equalTo(view.safeAreaLayoutGuide).inset(66)
-            $0.left.equalTo(view).inset(16)
+            $0.leading.equalTo(view).inset(16)
         }
         
         view.addSubview(closeButton)
         closeButton.snp.makeConstraints {
             $0.top.equalTo(view.safeAreaLayoutGuide).inset(9)
-            $0.right.equalTo(view)
+            $0.trailing.equalTo(view.safeAreaLayoutGuide)
             $0.width.height.equalTo(44)
         }
-
-        view.addSubview(createView)
-        createView.snp.makeConstraints {
-            $0.top.equalTo(titleLabel).inset(66)
-            $0.left.right.equalTo(view).inset(16)
-            $0.bottom.equalTo(view).inset(200)
-        }
-        
-        configureRoomsTitle()
-        
-        view.addSubview(personView)
-        personView.snp.makeConstraints {
-            $0.top.equalTo(titleLabel).inset(66)
-            $0.left.right.equalTo(view).inset(16)
-            $0.bottom.equalTo(view).inset(200)
-        }
-        
-        configureRoomsPerson()
-        
-        view.addSubview(dateView)
-        dateView.snp.makeConstraints {
-            $0.top.equalTo(titleLabel).inset(66)
-            $0.left.right.equalTo(view).inset(16)
-            $0.bottom.equalTo(view).inset(200)
-        }
-
-        configureRoomsDate()
         
         view.addSubview(nextButton)
         nextButton.snp.makeConstraints {
-            $0.left.right.equalTo(view).inset(16)
-            $0.bottom.equalTo(view).inset(57)
+            $0.leading.trailing.equalTo(view.safeAreaLayoutGuide).inset(16)
+            $0.bottom.equalTo(view.safeAreaLayoutGuide).inset(57)
             $0.height.equalTo(60)
         }
         
+        view.addSubview(nameView)
+        nameView.snp.makeConstraints {
+            $0.top.equalTo(titleLabel.snp.bottom).inset(-66)
+            $0.leading.trailing.equalToSuperview().inset(16)
+        $0.bottom.equalTo(nextButton.snp.top)
+        }
+
+        view.addSubview(personView)
+        personView.snp.makeConstraints {
+            $0.top.equalTo(titleLabel.snp.bottom).offset(66)
+            $0.leading.trailing.equalToSuperview().inset(16)
+            $0.bottom.equalTo(nextButton.snp.top)
+        }
+
+        view.addSubview(dateView)
+        dateView.snp.makeConstraints {
+            $0.top.equalTo(titleLabel.snp.bottom).offset(66)
+            $0.leading.trailing.equalToSuperview().inset(16)
+            $0.bottom.equalTo(nextButton.snp.top)
+        }
     }
     
+    // MARK: - Configure
     override func configUI() {
         super.configUI()
         view.backgroundColor = .backgroundGrey
     }
     
     // MARK: - Selectors
-    @objc func didTapDateBackView(_ gesture: UITapGestureRecognizer){
-        print("gesture")
-    }
-    
-    @objc func didSlider(_ slider: UISlider){
-        let value = slider.value
-        personLabel.text = "X \(Int(value))인"
-    }
-    
     @objc func didTapCloseButton(){
         print("ttaapp")
     }
     
     @objc func didTapnextButoon(){
         print("tap")
-    }
-    
-
-    // MARK: - Configure
-
-    func configureRoomsTitle(){
-
-        createView.addSubview(roomsNameTextField)
-        roomsNameTextField.snp.makeConstraints {
-            $0.top.equalTo(createView)
-            $0.left.right.equalTo(createView)
-            $0.height.equalTo(60)
-        }
-        
-        createView.addSubview(roomsTextLimit)
-        roomsTextLimit.snp.makeConstraints {
-            $0.top.equalTo(roomsNameTextField.snp.bottom).inset(-10)
-            $0.right.equalTo(createView)
-        }
-    }
-    
-    func configureRoomsPerson(){
-        
-        personView.addSubview(peronsViewLabel)
-        peronsViewLabel.snp.makeConstraints {
-            $0.top.equalTo(personView)
-            $0.left.right.equalTo(personView)
-        }
-        
-        personView.addSubview(personBackView)
-        personBackView.snp.makeConstraints {
-            $0.top.equalTo(peronsViewLabel.snp.bottom).inset(-36)
-            $0.left.right.equalTo(personView)
-            $0.height.equalTo(140)
-        }
-        
-        personBackView.addSubview(personLabel)
-        personLabel.snp.makeConstraints {
-            $0.centerX.equalToSuperview().offset(20)
-            $0.centerY.equalToSuperview()
-        }
-        
-        personBackView.addSubview(imageView)
-        imageView.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.right.equalTo(personLabel.snp.left)
-            $0.width.height.equalTo(60)
-        }
-        
-        personView.addSubview(personSlider)
-        personSlider.snp.makeConstraints {
-            $0.top.equalTo(personBackView.snp.bottom).inset(-54)
-            $0.left.equalTo(40)
-            $0.right.equalTo(-50)
-        }
-        
-        personView.addSubview(minLabel)
-        minLabel.snp.makeConstraints {
-            $0.right.equalTo(personSlider.snp.left).offset(-5)
-            $0.centerY.equalTo(personSlider.snp.centerY)
-        }
-        
-        personView.addSubview(maxLabel)
-        maxLabel.snp.makeConstraints {
-            $0.left.equalTo(personSlider.snp.right).offset(5)
-            $0.centerY.equalTo(personSlider.snp.centerY)
-        }
-    }
-    
-    func configureRoomsDate(){
-        
-        dateView.addSubview(dateViewLabel)
-        dateViewLabel.snp.makeConstraints {
-            $0.top.equalTo(dateView)
-            $0.left.right.equalTo(dateView)
-        }
-        
-        dateView.addSubview(dateBackView)
-        dateBackView.snp.makeConstraints {
-            $0.top.equalTo(peronsViewLabel.snp.bottom).inset(-36)
-            $0.left.right.equalTo(dateView)
-            $0.height.equalTo(60)
-        }
-        
-        dateBackView.addSubview(dateLabel)
-        dateLabel.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.centerY.equalToSuperview()
-        }
-        
-        let gestureTapRecognizer = UITapGestureRecognizer(target: self, action: #selector(didTapDateBackView(_:)))
-        gestureTapRecognizer.numberOfTapsRequired = 1
-        gestureTapRecognizer.numberOfTouchesRequired = 1
-        dateView.addGestureRecognizer(gestureTapRecognizer)
-
     }
     
     // MARK: - Navigation
